+++ shared/types/src/hasura/contributions.ts
@@ -21,6 +21,7 @@ export type ContributionsAnswers = {
   agreement: ContributionAgreement;
   updatedAt: string;
   statuses?: ContributionStatus[];
+  display_date?: string;
 };
 
 export type ContributionQuestion = {
