+++ targets/frontend/package.json
@@ -11,6 +11,7 @@
     "@hookform/resolvers": "^3.3.1",
     "@mui/icons-material": "5.11.16",
     "@mui/material": "5.14.11",
+    "@mui/x-date-pickers": "^7.13.0",
     "@reach/accordion": "^0.16.1",
     "@reach/dialog": "^0.16.0",
     "@reach/menu-button": "^0.16.1",
