+++ targets/frontend/src/components/contributions/answers/answer.mutation.ts
@@ -57,6 +57,7 @@ export type MutationProps = Pick<
   | "legiReferences"
   | "otherReferences"
   | "cdtnReferences"
+  | "displayDate"
 > & {
   status: string;
   userId: string;
