+++ targets/frontend/src/components/forms/index.ts
@@ -7,3 +7,4 @@ export * from "./ToggleButtonGroup";
 export * from "./File";
 export * from "./AutocompleteChips";
 export * from "./Switch";
+export * from "./DatePicker";
