+++ targets/frontend/src/modules/contribution/mapContributionToDocument.ts
@@ -8,7 +8,7 @@ import { getReferences } from "./getReferences";
 import { generateCdtnId } from "@shared/utils";
 import { generateContributionSlug } from "./generateSlug";
 import { getGlossaryContent } from "../common/getGlossaryContent";
-import { format } from "date-fns";
+import { format, parseISO } from "date-fns";
 
 async function getBaseDocument(
   data: ContributionsAnswers,
@@ -78,7 +78,11 @@ export const mapContributionToDocument = async (
 
   const initalDoc: ContributionDocumentJson = {
     ...baseDoc,
-    date: format(new Date(data.updatedAt), "dd/MM/yyyy"),
+    date: data.display_date
+      ? format(parseISO(data.display_date), "dd/MM/yyyy")
+      : data.updatedAt
+      ? format(parseISO(data.updatedAt), "dd/MM/yyyy")
+      : undefined,
     contentType: data.content_type,
     linkedContent: data.cdtn_references.map((v) => ({
       cdtnId: v.cdtn_id!,
