+++ targets/frontend/src/modules/documents/api/documents.service.ts
@@ -1,22 +1,18 @@
 import { DocumentsRepository } from "./documents.repository";
 import { ConflictError, NotFoundError } from "src/lib/api/ApiErrors";
-import { Information, InformationsRepository } from "src/modules/informations";
-import { format, parseISO } from "date-fns";
-import { generateCdtnId, generateInitialId } from "@shared/utils";
-import slugify from "@socialgouv/cdtn-slugify";
+import { InformationsRepository } from "src/modules/informations";
 import {
   ContributionRepository,
   mapContributionToDocument,
 } from "src/modules/contribution";
 import { ModelRepository } from "../../models/api";
-import { Model } from "../../models";
-import { AgreementDoc, HasuraDocument } from "@socialgouv/cdtn-types";
 import { generateContributionSlug } from "src/modules/contribution/generateSlug";
 import { AgreementRepository } from "../../agreements/api";
-import { Agreement } from "../../agreements";
 import { SourceRoute } from "@socialgouv/cdtn-sources";
-import { getGlossaryContent } from "src/modules/common/getGlossaryContent";
 import pMap from "p-map";
+import { mapAgreementToDocument } from "src/modules/agreements/mapAgreementToDocument";
+import { mapInformationToDocument } from "src/modules/informations/mapInformationToDocument";
+import { mapModelToDocument } from "src/modules/models/mapModelToDocument";
 
 export class DocumentsService {
   private readonly informationsRepository: InformationsRepository;
@@ -39,187 +35,6 @@ export class DocumentsService {
     this.agreementRepository = agreementRepository;
   }
 
-  private async mapInformationToDocument(
-    data: Information,
-    document?: HasuraDocument<any>
-  ): Promise<HasuraDocument<any>> {
-    const introWithGlossary = await getGlossaryContent(
-      "markdown",
-      data.intro ?? ""
-    );
-    return {
-      cdtn_id: document?.cdtn_id ?? generateCdtnId(data.title),
-      initial_id: data.id ?? generateInitialId(),
-      source: "information",
-      meta_description: data.metaDescription ?? data.description,
-      title: data.title,
-      text: data.title,
-      slug: document?.slug ?? slugify(data.title),
-      is_searchable: document ? document.is_searchable : true,
-      is_published: document ? document.is_published : true,
-      is_available: true,
-      document: {
-        date: data.updatedAt
-          ? format(new Date(data.updatedAt), "dd/MM/yyyy")
-          : undefined,
-        intro: data.intro,
-        introWithGlossary,
-        description: data.description,
-        sectionDisplayMode: data.sectionDisplayMode,
-        dismissalProcess: data.dismissalProcess,
-        references: data.references.length
-          ? [
-              {
-                label: data.referenceLabel,
-                links: data.references,
-              },
-            ]
-          : undefined,
-        contents: await Promise.all(
-          data.contents.map(
-            async ({ name, title, blocks, references, referenceLabel }) => {
-              return {
-                name,
-                title,
-                blocks: await Promise.all(
-                  blocks.map(async (block) => {
-                    const htmlWithGlossary = await getGlossaryContent(
-                      "markdown",
-                      block.content ?? ""
-                    );
-                    return {
-                      type: block.type,
-                      ...(block.type === "content"
-                        ? {
-                            title: block.content,
-                          }
-                        : {
-                            markdown: block.content,
-                            htmlWithGlossary,
-                          }),
-                      ...(block.type === "graphic"
-                        ? {
-                            size: block.file?.size,
-                            imgUrl: block.img?.url,
-                            altText: block.img?.altText,
-                            fileUrl: block.file?.url,
-                          }
-                        : {}),
-                      ...(block.type === "content"
-                        ? {
-                            blockDisplayMode: block.contentDisplayMode,
-                            contents: block.contents?.length
-                              ? block.contents.map(({ document }) => {
-                                  return {
-                                    title: document.title,
-                                    cdtnId: document.cdtnId,
-                                    source: document.source,
-                                  };
-                                })
-                              : undefined,
-                          }
-                        : {}),
-                    };
-                  })
-                ),
-                references: references?.length
-                  ? [
-                      {
-                        label: referenceLabel,
-                        links: references,
-                      },
-                    ]
-                  : undefined,
-              };
-            }
-          )
-        ),
-      },
-    };
-  }
-
-  private mapModelToDocument(
-    data: Model,
-    document?: HasuraDocument<any>
-  ): HasuraDocument<any> {
-    return {
-      cdtn_id: document?.cdtn_id ?? generateCdtnId(data.title),
-      initial_id: data.id!,
-      source: "modeles_de_courriers",
-      meta_description: data.metaDescription,
-      title: data.title,
-      text: data.intro,
-      slug: document?.slug ?? slugify(data.title),
-      is_searchable: document ? document.is_searchable : true,
-      is_published: document ? document.is_published : true,
-      is_available: true,
-      document: {
-        meta_title: data.metaTitle,
-        type: data.type,
-        date: format(parseISO(data.updatedAt), "dd/MM/yyyy"),
-        author: "Ministère du Travail",
-        references: data.legiReferences
-          .map((item) => ({
-            url: `https://www.legifrance.gouv.fr/codes/article_lc/${item.legiArticle.cid}`,
-            title: item.legiArticle.label,
-            type: "external",
-          }))
-          .concat(
-            data.otherReferences.map((item) => ({
-              url: item.url ?? "",
-              title: item.label,
-              type: "external",
-            }))
-          ),
-        description: data.metaDescription,
-        intro: data.intro,
-        filename: data.file.url,
-        filesize: parseInt(data.file.size ?? "0"),
-        html: data.previewHTML,
-      },
-    };
-  }
-
-  private mapAgreementToDocument(
-    data: Agreement,
-    document?: HasuraDocument<AgreementDoc>
-  ): HasuraDocument<AgreementDoc> {
-    return {
-      cdtn_id: document?.cdtn_id ?? generateCdtnId(data.name),
-      initial_id: data.id,
-      source: "conventions_collectives",
-      meta_description: `IDCC ${data.id}: ${data.name}`,
-      title: data.name,
-      text: `IDCC ${data.id}: ${data.name} ${data.shortName}`,
-      slug: document?.slug ?? slugify(data.shortName),
-      is_searchable: document
-        ? document.is_searchable
-        : data.kali_id !== undefined,
-      is_available: true,
-      is_published: data.kali_id !== undefined,
-      document:
-        data.kali_id !== undefined
-          ? {
-              date_publi: data.publicationDate
-                ? `${data.publicationDate}T00:00:00.000Z`
-                : undefined,
-              effectif: data.workerNumber ?? undefined,
-              num: Number(data.id),
-              url: data.legifranceUrl ?? undefined,
-              shortTitle: data.shortName,
-              synonymes: data.synonyms,
-            }
-          : {
-              date_publi: data.publicationDate
-                ? parseISO(data.publicationDate).toISOString()
-                : undefined,
-              num: Number(data.id),
-              shortTitle: data.shortName,
-              synonymes: data.synonyms,
-            },
-    };
-  }
-
   public async publish(id: string, source: SourceRoute) {
     let document = await this.documentsRepository.fetch({
       source,
@@ -237,7 +52,7 @@ export class DocumentsService {
             cause: null,
           });
         }
-        document = await this.mapInformationToDocument(information, document);
+        document = await mapInformationToDocument(information, document);
         break;
       case "contributions":
         const contribution = await this.contributionRepository.fetch(id);
@@ -284,7 +99,7 @@ export class DocumentsService {
             cause: null,
           });
         }
-        document = this.mapModelToDocument(model, document);
+        document = mapModelToDocument(model, document);
         break;
 
       case "conventions_collectives":
@@ -296,7 +111,7 @@ export class DocumentsService {
             cause: null,
           });
         }
-        document = this.mapAgreementToDocument(agreement, document);
+        document = mapAgreementToDocument(agreement, document);
         break;
 
       default:
