+++ targets/frontend/src/modules/informations/api/informations.query.ts
@@ -13,6 +13,7 @@ export const informationsQuery = gql`
       sectionDisplayMode
       title
       updatedAt
+      displayDate
       dismissalProcess
       contents(order_by: { order: asc }) {
         id
