+++ targets/frontend/src/modules/informations/components/informationsEdit/editInformation.mutation.ts
@@ -50,6 +50,7 @@ export const informationMutation = gql`
           metaDescription
           referenceLabel
           sectionDisplayMode
+          displayDate
         ]
       }
     ) {
