+++ targets/frontend/src/modules/informations/type.ts
@@ -94,5 +94,6 @@ export const informationSchema = z.object({
   updatedAt: z.string().nullable().optional(),
   contents: z.array(informationContentSchema),
   references: z.array(referenceSchema),
+  displayDate: z.string().nullable().optional(),
 });
 export type Information = z.infer<typeof informationSchema>;
