+++ targets/frontend/src/modules/models/components/graphql.type.ts
@@ -60,4 +60,5 @@ export type ModelModelsInsertInput = {
   title?: string;
   type?: string;
   updatedAt?: string;
+  displayDate?: string;
 };
