+++ targets/frontend/src/modules/models/type.ts
@@ -32,6 +32,7 @@ export const modelSchema = z.object({
   file: fileSchema,
   legiReferences: z.array(legiReferenceSchema),
   otherReferences: z.array(otherReferenceSchema),
+  displayDate: z.string(),
 });
 
 export type Model = z.infer<typeof modelSchema>;
